consolidate useShapeInternal into useShape, rename hooks to useShapeSingle/useShapeList, fix async noop
